fix : 7단계 - ArticleRepository에 새로 작성한 메서드를 수정
- 이름을 수정하고 로직에서 잘못된 부분을 수정

+++ src/main/java/db/ArticleRepository.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface ArticleRepository {
     Article save(Article article);
-    List<Article> findTopNLessThanByIdDecreasingOrder(int limit, int id);
+    List<Article> findTopNLessThanByIdDecreasingOrder(int limit, long id);
 }
